# docs/swagger.yaml
basePath: /api/v1/wallet
definitions:
...
  resp.BasePageResp:
    properties:
      count:
        description: 数量
        type: integer
      currentPage:
        description: 当前页数
        type: integer
      pageSize:
        description: 每页数据条数
        type: integer
      totalPage:
        description: 总页数
        type: integer
    type: object
...
  resp.CurrencyQuoteInfoResp:
    properties:
      base_currency:
        description: 基础货币
        type: string
      code:
        description: 标示
        type: string
      id:
        description: id
        type: integer
      price:
        description: 金额
        type: number
      quote_currency:
        description: 基础货币
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  resp.CurrencyQuoteListResp:
    properties:
      items:
        description: 帐号详情
        items:
          $ref: '#/definitions/resp.CurrencyQuoteInfoResp'
        type: array
      page:
        $ref: '#/definitions/resp.BasePageResp'
        description: 分页
        type: object
    type: object
  resp.ReadAllDepositDetailInfoResp:
    properties:
      name:
        description: 用户名
        type: string
      order_id:
        description: 订单id
        type: integer
      source:
        description: 来源 0：充值1：后台充值
        type: integer
      status:
        description: 状态 0确认中,1已确定
        type: integer
      symbol:
        description: 代币代号
        type: string
      transaction_hash:
        description: hash
        type: string
      type:
        description: 标识 coin,token
        type: string
      uid:
        description: 用户id
        type: string
      updated_at:
        description: 更新时间
        type: string
      value:
        description: 金额
        type: number
    type: object
  resp.ReadAllDepositDetailResp:
    properties:
      items:
        description: 数据
        items:
          $ref: '#/definitions/resp.ReadAllDepositDetailInfoResp'
        type: array
      page:
        $ref: '#/definitions/resp.BasePageResp'
        description: 分页
        type: object
    type: object
...
info:
  contact: {}
  description: 测试
  license: {}
  title: 测试
  version: 0.0.1
paths:
...
  /admin/currency/quote/create:
    post:
      description: 创建货币汇率
      parameters:
      - description: 基础货币symbol
        in: formData
        name: base_currency
        required: true
        type: string
      - description: 报价货币symbol
        in: formData
        name: quote_currency
        required: true
        type: string
      - description: 金额
        in: formData
        name: price
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 创建货币汇率接口
      tags:
      - Account 后台钱包-用户钱包
  /admin/currency/quote/list:
    get:
      description: 获取汇率分页
      parameters:
      - description: 长度
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 搜索标示
        in: query
        name: keyword
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: integer
      - description: 结束时间
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CurrencyQuoteListResp'
      security:
      - ApiKeyAuth: []
      summary: 获取汇率分页接口
      tags:
      - Account 后台钱包-用户钱包
  /admin/currency/quote/update:
    post:
      description: 更新货币汇率
      parameters:
      - description: 汇率id
        in: formData
        name: id
        required: true
        type: integer
      - description: 金额
        in: formData
        name: price
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新货币汇率接口
      tags:
      - Account 后台钱包-用户钱包
  /admin/currency/quote_history/list:
    get:
      description: 获取汇率历史记录分页
      parameters:
      - description: 长度
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 搜索标示
        in: query
        name: keyword
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: integer
      - description: 结束时间
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CurrencyQuoteListResp'
      security:
      - ApiKeyAuth: []
      summary: 获取汇率历史记录分页接口
      tags:
      - Account 后台钱包-用户钱包
  /admin/operate/recharge/audio:
    post:
      description: 删除充值
      parameters:
      - description: 充值id
        in: formData
        name: id
        required: true
        type: string
      - description: 财务id
        in: formData
        name: financial_id
        required: true
        type: integer
      - description: 状态1：通过；2：不通过
        in: formData
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除充值接口
      tags:
      - Account 后台运营-充值
  /admin/operate/recharge/join:
    post:
      description: 添加充值
      parameters:
      - description: 用户id
        in: formData
        name: uid
        required: true
        type: string
      - description: 币种标示
        in: formData
        name: symbol
        required: true
        type: string
      - description: 金额
        in: formData
        name: money
        type: number
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 添加充值接口
      tags:
      - Account 后台运营-充值
  /admin/operate/recharge/list:
    get:
      description: 充值列表
      parameters:
      - description: 长度
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 搜索用户
        in: query
        name: keyword
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: integer
      - description: 结束时间
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadAllDepositDetailResp'
      security:
      - ApiKeyAuth: []
      summary: 充值列表接口
      tags:
      - Account 后台运营-充值
  /admin/operate/recharge/remove:
    post:
      description: 删除充值
      parameters:
      - description: 充值id
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除充值接口
      tags:
      - Account 后台运营-充值
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"